OpenSiv3D v4.1.2に更新、EasyViewerの修正

+++ FacilityState.cpp
@@ -16,7 +16,7 @@ void FacilityState::draw(const BasicCamera3D& camera)
 
 	auto k = EaseOutQuart(m_constructionProgress);
 	auto s = (9.0 + Sin(m_progress * Math::TwoPi)) / 10.0;
-	Mat4x4 mat = Mat4x4::Scaling(scale / s, scale * k * s * s * (1.0 - g_planetManagerPtr->m_destroy), scale / s)
+	Mat4x4 mat = Mat4x4::Scaling(float(scale / s), float(scale * k * s * s * (1.0 - g_planetManagerPtr->m_destroy)), float(scale / s))
 		* Mat4x4::Translation(0, float(m_region.lock()->getPosition().length()), 0)
 		* Mat4x4::RotationZ(-f) * Mat4x4::RotationY(t)
 		* camera.getMat4x4();
